Major overhaul to events retrieval to fix recurrences.
This is necessary to conform to the icalendar and caldav view of recurring events as well as to fix broken code. The main conceptual change is to group events with the same name as a single resource. Thus a retrieval by name will result in a single event with associated overrides (and possibly instances if expansion is requested).

Retrieval by guid can result in multiple events (from the inbox for example) but they will be grouped by name as above.

This corresponds to the icalendar approach of delivering a master event + overrides. 

A new parameter class RecurringRetrievalMode is used to indicate how recurring events should be handled. CalDAV allows for master + overrides or for a full expansion with all times in UTC. Bedework also allows full expansion retaining the timezone information. In additon, expansions and overides may be limited by a time period different to that of the search.

Changes are extensive. Calls on getEvent, getEvents and others renamed to those names are affected as is interpretation of the results. For the most part, the web client code is unchanged other than to change parameters so events are retrieved expanded with timezones. Downloading differs slightly.

The main visible change to the results is the chaining together of overrides and instances through the EventInfo class.

As a result of this some hibernate problems resurfaced requiring an overhaul mainly to the web client code. Action classes invoked by action urls have been changed to set information allowing render url invoked classes to do the actual retrieval. It should be possible to reenable much of the lazy loading disabled earlier. These changes did affect the web client more than the back end code.

Updating of recurring events via CalDAV has also been partially fixed. Some problems remain. Currently (as before) updating of recurring events in the web client is disabled.