Allow subscriptions to web calendars via the CalDAV server.
This requires an extra element in the options file to specify the url fro the web calendar service.

This update should allow Apple ical users to subscribe to sahred calendars as an interim measure until we deploy bedework aliases.

It also should be usable by Google etc. This does not implement the filtering which the user and public events interface have. These features will probably appear as filtered subscriptiuons in bedework.