Implement resources folders which can be created within a calendar collection. This follows the approach outlined in RFC 4791 and will provide a place to store shareable attachments while ensuring access restrictions are applied.
In addition, implement support for creating and retrieving 'file' resources which may be stored in any non-calendar collection. This will probably be useful when creating an image repository for example, but is also required for full implementation of shared attachments.

Currently, there is no ui support fo any of this, files may be uploaded with DAV explorer and viewed via a browser accessing the caldav server.