This update makes significant changes to scheduling to bring it closer to the (unreleased) draft 7 of CalDAV scheduling.
This does cause a schema change. See the end of the message for commands to update hsql - or just recreate.

So far basic non-recurring scheduling between an organizer and attendee mostly works. 

Currently the inbox is being updated and to some extent is getting in the way. UI and back end changes are coming.

Updates from the organizer are not being processed yet - soon will.

    create table bw_sched_msg (
        id integer generated by default as identity (start with 1),
        bwseq integer not null,
        bw_timestamp varchar(16) not null,
        bw_sequence integer,
        bw_lastproc varchar(16) not null,
        bw_inbox char(1) not null,
        bw_principal_href varchar(255),
        bw_eventname varchar(500),
        bw_recurrence_id varchar(16),
        primary key (id)
    );

    create index bwidx_smsg_eventname on bw_sched_msg (bw_eventname);

    create index bwidx_smsg_timestamp on bw_sched_msg (bw_timestamp);

    create index bwidx_smsg_principal on bw_sched_msg (bw_principal_href);