Fix an issue with recurring events in which attendees are invited to instances only.
Missing columns in the annotations schema so using x-props to store the flags for the time being.

This finishes off the move to drop use of PropertyInfo and use BwIcalPropertyInfo instead - generated from annotations.

Update annotation so that it can be used on getters and setters - also add information that was in PropertyInfo.

Minor change to reduce visibility of a method in CalSvc