Two intertwined updates here.
Most projects are changed to go some way towards maven. Use maven-ant tasks to deal with build dependencies. Mostly chnages to build.xml and the additon of pom.xml.

Additionally, added a new bedenote project which handles JMS messages. This removes the need for special support from activemq.

Removed the auto-deploy of activemq config and made it a special target - "deployActivemq"

To install this stuff:

1. Check out the bedenote project
2. execute the deployActivemq target
3. Build/deploy bedenote
4. Restart