Small change to api for move Event to pass the entire calendar objects for event moves
Drop unused groupCalendarRootPath in CoreCalendars
Use buildPath for userCalendarRootPath - and fix it so it works.

Add more information to system events - also add an interface for notification carrying events.

Had to add a shared parameter to many internal methods. This may change.

Getting closer to adding notifications to a users collection.

Added a method to BwCalendar to set sharing on/off

Also a property to indicate if notifications are wanted. Added a system default.